Switch lan78xx into polling mode
The problem was related to some taints when the phy cable of the
ethernet was plugged in and out, leading to the driver that
does not detect phy cable in/out any more.
This hack/modification puts the driver into polling mode which is
also supported by the driver.

This also causes the kernel to go into deadlock on akita.
On kernel 5.15 there are additional changes in the phylib where threaded interrupts does not work properly